Renamed Solution from Epi.Extensions.Settings to AddOn.Episerver.Settings
