+++ README.md
@@ -1,4 +1,4 @@
-# Epi.Extensions.Settings
+# AddOn.Episerver.Settings
 
 [![Platform](https://img.shields.io/badge/platform-.NET%204.6.1-blue.svg?style=flat)](https://msdn.microsoft.com/en-us/library/w0x726c2%28v=vs.110%29.aspx)
 [![Platform](https://img.shields.io/badge/EPiServer-%2011.1.0-orange.svg?style=flat)](http://world.episerver.com/cms/)
