+++ src/AddOn.Episerver.Settings/AddOn.Episerver.Settings.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{D999E2CF-F01E-48E6-85ED-98AEEE83B6FB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Epi.Extensions.Settings</RootNamespace>
-    <AssemblyName>Epi.Extensions.Settings</AssemblyName>
+    <RootNamespace>AddOn.Episerver.Settings</RootNamespace>
+    <AssemblyName>AddOn.Episerver.Settings</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
