+++ src/AddOn.Episerver.Settings/Core/SettingsContentTypeAttribute.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="GlobalSettingsComponent.cs" company="none">
+// <copyright file="SettingsContentTypeAttribute.cs" company="none">
 //      Copyright © 2019 Linus Ekström, Jeroen Stemerdink.
 //      Permission is hereby granted, free of charge, to any person obtaining a copy
 //      of this software and associated documentation files (the "Software"), to deal
@@ -21,28 +21,30 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace Epi.Extensions.Settings.UI
+namespace AddOn.Episerver.Settings.Core
 {
-    using EPiServer.Shell.ViewComposition;
+    using System;
+
+    using EPiServer.DataAnnotations;
 
     /// <summary>
-    /// Component that provides a tree based navigation for global settings.
-    /// Implements the <see cref="EPiServer.Shell.ViewComposition.ComponentDefinitionBase" />
+    /// Class SettingsContentTypeAttribute. This class cannot be inherited.
+    /// Implements the <see cref="EPiServer.DataAnnotations.ContentTypeAttribute" />
     /// </summary>
-    /// <seealso cref="EPiServer.Shell.ViewComposition.ComponentDefinitionBase" />
-    [Component]
-    public sealed class GlobalSettingsComponent : ComponentDefinitionBase
+    /// <seealso cref="EPiServer.DataAnnotations.ContentTypeAttribute" />
+    [AttributeUsage(validOn: AttributeTargets.Class)]
+    public sealed class SettingsContentTypeAttribute : ContentTypeAttribute
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GlobalSettingsComponent"/> class.
+        /// Gets or sets the settings instance unique identifier.
         /// </summary>
-        public GlobalSettingsComponent()
-            : base("epi-cms/component/MainNavigationComponent")
-        {
-            this.LanguagePath = "/episerver/cms/components/globalsettings";
-            this.Title = "Global settings";
-            this.SortOrder = 100;
-            this.Settings.Add(new Setting("repositoryKey", value: GlobalSettingsRepositoryDescriptor.RepositoryKey));
-        }
+        /// <value>The settings instance unique identifier.</value>
+        public string SettingsInstanceGuid { get; set; }
+
+        /// <summary>
+        /// Gets or sets the name of the settings.
+        /// </summary>
+        /// <value>The name of the settings.</value>
+        public string SettingsName { get; set; }
     }
 }
