+++ src/AddOn.Episerver.Settings/LICENSE
