+++ src/AddOn.Episerver.Settings/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Epi.Extensions.Settings")]
+[assembly: AssemblyTitle("AddOn.Episerver.Settings")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Epi.Extensions.Settings")]
-[assembly: AssemblyCopyright("Copyright © 2019 Linus Ekström, Jeroen Stemerdink")]
+[assembly: AssemblyProduct("AddOn.Episerver.Settings")]
+[assembly: AssemblyCopyright("Copyright © 2020 Linus Ekström, Jeroen Stemerdink")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -33,5 +33,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.1.0")]
-[assembly: AssemblyFileVersion("1.0.1.0")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
