+++ src/AddOn.Episerver.Settings/Resources/Epi.Libraries.Settings.EmbeddedLangFiles_DA.xml
