+++ src/AddOn.Episerver.Settings/Resources/Epi.Libraries.Settings.EmbeddedLangFiles_DE.xml
