+++ src/AddOn.Episerver.Settings/UI/SettingsRepositoryDescriptor.cs
@@ -21,12 +21,12 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace Epi.Extensions.Settings.UI
+namespace AddOn.Episerver.Settings.UI
 {
     using System;
     using System.Collections.Generic;
 
-    using Epi.Extensions.Settings.Core;
+    using AddOn.Episerver.Settings.Core;
 
     using EPiServer.Cms.Shell.UI.CompositeViews.Internal;
     using EPiServer.Core;
