+++ src/AddOn.Episerver.Settings/UI/SettingsView.cs
@@ -21,11 +21,11 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace Epi.Extensions.Settings.UI
+namespace AddOn.Episerver.Settings.UI
 {
     using System.Collections.Generic;
 
-    using Epi.Extensions.Settings.Core;
+    using AddOn.Episerver.Settings.Core;
 
     using EPiServer.Core;
     using EPiServer.Framework.Localization;
