+++ src/AddOn.Episerver.Settings/module.config
