+++ src/AddOn.Episerver.Settings/packages.config
